Migrate StreamType enum over to api::v2::updates module
The StreamType is among the last bits of functionality that remains from
the legacy event streaming bits. This enum is still useful in the
context of the subscription based streaming and so we move it into the
api::v2::updates module. Note, however, that it is entirely an
implementation detail and no longer exposed publicly.