Move websocket connect functionality into websocket module
This change concludes the rework of our websocket streaming
functionality. It moves the truly last bit of the events module -- the
connect functionality -- into a new module in the root of the crate,
websocket.
Note that in the past we exposed certain low-level details of the events
module publicly, in order to allow clients to stream raw (i.e.,
undecoded) updates, among other things. Currently we do not have a
replacement available for this functionality, but it is possible that
one will be added in the future.