Remove internal stream related types/methods from the documentation
Unfortunately some of our types internally used in the stream module
appear in trait definitions on public types. Rust enforces that these
former types be declared public. However, users ultimately should never
have to work with them. As such, having them appear in the generated
documentation can be a source of confusion.
With this change we hide these types from the documentation, so as to
prevent such confusion from the get-go. We do the same for the as_str
method on the Source trait, as it only has relevance to internal code
and users cannot prevent this trait, for it is sealed.