Make Subscribable futures implement Send
We have tagged the async Subscribable trait as ?Send, resulting in the
futures returned by its method not implementing the Send trait. There is
actually no really good reason for that and it may significantly
complicate downstream usage.
With this change we adjust the trait to make sure that all returned
futures are Send.