Disable invalid credentials test
Without notice or version bump, Alpaca has decided to change the HTTP
status code reported when providing invalid credentials from 401 to 403.
This status code is now in conflict with that of specific endpoints.
E.g., the orders endpoint reports this error also when submitting an
order for more funds than the account is eligible to trade. The only way
to differentiate between the two conditions is to look at the "code"
being reported, which of course is not documented anywhere and,
presumably, can equally be changed on a whim -- the reason why we do not
currently define these anywhere ourselves.
The only solution to this problem at this point is to ignore it: we
disable the test that exercises this path. Clients will need to
potentially deal with "unknown" errors (when making an unauthenticated
request to an endpoint not supporting a 403 error) or differentiate
themselves (for endpoints that carry a 403 error definition).