Simplify query creation for QuotesReq
The creation of the query for the QuotesReq type is a bit cumbersome,
because we create our own internal type just to exclude the symbol from
being included in the resulting query string. There is actually a
simpler way to accomplish the same feat: just tag the corresponding
member variable with #[serde(skip)].
With this change we do exactly that and simplify the code.