Use usize for specifying limits
A bunch of the Alpaca endpoints accept some form of limit parameters.
Right now we specify some of them as u64 and others as usize. None of
these limits is really running risk of exceeding the range of the usize
type on any of the targets that we care about or expect the crate to be
used on. As such, using usize seems like the more general choice and
with this change we switch everything over to using usize.