Add support for streaming realtime quotes
This change adds support for streaming realtime market data quotes to
the crate. As the most user facing part of the change, that means that
clients can now set quotes to subscribe to on
data::v2::stream::MarketData objects and may now receive the new
Data::Quote variant which contains quote data.