Introduce api::v2::order::Status::is_terminal method
Alpaca more or less implicitly defines a couple of order statuses as
terminal, stating that they will not change once reached. Having such
information easier accessible to clients makes sense, and so with this
change we introduce the Status::is_terminal method, that checks whether
such a status is terminal.