Make crypto asset class known
Some time back, Alpaca added support for trading of crypto currencies.
The way they are represented is via a new asset class. We do not have
immediate plans of adding full crypto support, but currently we fail
deserialization of inquiries of all open positions when some of them are
for crypto currencies, because our Position object contains an
associated asset class.
With this change we make this asset class known.