Make api::v2::asset::Class non-exhaustive
We do not currently have a formalized method for when we tag an enum or
struct as non-exhaustive, though we do have some kind of system that we
try to follow.
In general it would probably be fair to say that the introduction of a
new asset class should not break existing code in the grand scheme of
things: if an asset class was previously unknown it would be okay to
simply ignore it. With that reasoning in mind, this change declares
api::v2::asset::Class as non-exhaustive, so that we can add new asset
classes without a breaking change.