Add Otc variant to api::v2::asset::Exchange enum
At some point Alpaca has introduced an OTC "exchange" for assets that
are traded over the counter. With this change we introduce a new variant
to the api::v2::asset::Exchange enum for this purpose. Furthermore, we
make said enum non-exhaustive, so that additions to it are no longer
considered breaking changes.