Remove From<Into<String>> conversion for asset::Symbol
The infallible conversion from anything that can be turned into a string
into an api::v2::asset::Symbol object is arguably too general, to the
point that it is misleading. Specifically, we support various symbol
variants, but we do not actually perform any parsing as part of the
aforementioned conversion. Rather, we just treat the "thing" that can be
turned into a string as the Symbol::Sym variant.That can easily and
unintentionally lead to issues where a user provided a different variant
(e.g., one of a form including an exchange such as SPY:NYSE), but this
string is just treated as a simple symbol.
With this change we remove this conversion. A future change will
introduce a fallible one, that actually parses the provided string,
which should be saner to use.