Make slice_to_str work with Cow
The slice_to_str accepts a function that returns a string reference with
a static lifetime. That is fine for our existing use cases, but it will
not be sufficient once we want to perform some kind of transformation.
To make this function a bit more flexible, this change adjust the
function to return a copy-on-write string instead, which gives us the
option of returning an owned object.