Make members of ApiInfo structure publicly accessible
In the past we were reluctant to make members of the ApiInfo type
publicly accessible. At least part of the reason was that we wanted to
avoid embedding yet another type -- Url -- from a different crate into
our public API surface.
However, we already embed url's error type in the public API.
Furthermore, with more URLs being embedded into ApiInfo, usability
becomes a concern if we do not provide more reasonable access of
individual members.
With this change we make all members of the ApiInfo publicly accessible.
We also tag the type as non_exhaustive, to effectively force users to
use one of the provided constructors and to be a bit more future-proof.