Rename ENV_API_URL constant to ENV_API_BASE_URL
We use the term "base URL" to convey the fact that a URL may not be
fully defining an endpoint. Most likely, that is the case because a path
is still missing. The nomenclature is exposed publicly, but the internal
ENV_API_URL constant does not yet adhere to it. With this change we
rename the constant to make its character more pronounced.