Deserialize empty Order::class string as Class::default()
We have seen instances where the class string of order objects is simply
reported as empty by the Alpaca API. That is obviously not a valid
variant.
While hopefully this issue will eventually be fixed by Alpaca, for now
we fix deserialization failures caused by that by mapping the empty
string to the default order class (simple).

Fixes #37