Introduce Trading API websocket URL to ApiInfo
While only necessary on rare occasions, we want to provide users with a
way to configure the URL for the Trading API's websocket update stream
to use.
With this change we introduce a new member to the ApiInfo type that
holds this URL. By default it is inferred based on the Trading API base
URL, which was also what we used in the past.