Rename DATA_WEBSOCKET_BASE_URL
We have decided to use "stream" as the terminology of choice for the
websocket streaming APIs, as the fact that a websocket connection is
used is more or less an implementation detail.
With this change we rename the DATA_WEBSOCKET_BASE_URL constant to
DATA_STREAM_BASE_URL for this very reason.