Make various types exhaustive again
As it turns out we cannot create a struct that is tagged as
non-exhaustive by initializing specifying values for all its members.
While that makes sense, that fact was not taken into consideration when
we applied this tag to various types. Given that we want to instantiate
objects of various types outside of this crate, this change removes the
non_exhaustive tag from the types in question.