Implement Deserialize trait for api::v2::order::TakeProfit
This change implements the Deserialize trait for the
api::v2::order::TakeProfit enum. We need this type to be deserializable
in order to eventually make order request objects deserializable as
well.
The way we make that possible somewhat easily is by converting
TakeProfit objects from a struct that we annotate with serde attributes
that allow us to deserialize from the expected state. We do the same for
serialization support, replacing the existing implementation of the
Serialize trait by an approach that converts into the trivially
serializable struct first.