Replace "trade update" terminology with "order update"
We use the term "trade update" in a bunch of places. That is basically
the terminology that Alpaca uses and we inherited it. But that is
arguably a source of confusion: the API does not use the term "trade"
anywhere else. In fact, trade updates are really order updates, as
evidenced by the fact that an Order object is embedded in the
corresponding message.
To convey the intentions better, this change converts the crate to use
"order update" as the new terminology.