Change api::v2::updates::Request to contain Cow members
The api::v2::updates::Request type currently contains two string
references. We actually want to be able to instantiate this type with
owned objects as well and so with this change we make it work with Cow
objects instead of pure references instead. That is in line with what we
already have for the Stream type in the same module.
On top of that, we also make the Request type deserializable, by auto
deriving the Deserialize trait for it.