Rename ErrorMessage to ApiError and export it
This change renames the ErrorMessage type to ApiError, which is found to
be a more suitable name. Furthermore, it turns out that we were missing
exporting this type publicly and so we fix that up. Note that members
were always accessible already, but the type itself could just not be
named -- complicating usage unnecessarily.