Implement Deref for data::v2::stream::Normalized
For the most part, instances of our data::v2::stream::Normalized type
are client constructed and then passed to subscribe to or unsubscribe
from real time data. However, there is an argument to be made that they
should still be inspectable by client code as well -- something not
currently possible.
With this change we enable such introspection by implementing the Deref
trait for the type, forwarding operations to the underlying slice.