Implement Deserialize for data::v2::stream::Request
There are a few occasions in which it may be necessary to deserialize a
data::v2::stream::Request object, instead of serializing it, as is much
more common.
To enable those use cases, this change implements the Deserialize trait
for it. We also make it publicly accessible, although undocumented, as
we did for the various order update types in the past.