Adjust stock split & spin off account activity strings
In typical Alpaca fashion, the strings used for representing the stock
split and spin-off account activity types have been renamed in a
backwards compatibility breaking manner and without any announcement.
This change adjusts the strings accordingly, hoping that the circus does
not get a new lead clown that pushes for numeric representation of this
crucial data in the next four months.