Add price member to api::v2::account_activities::NonTradeActivity
It appears that non-trade activities, which the documentation clearly
states are anything but FILLs, can also have a price associated with
them. We have certainly seen that being the case for stock splits,
despite the documentation not making any mention of it.
In order to make this member accessible in a structured way and cutting
down on the need to parse the associated description, this change a
corresponding optional member to the type.