Adjust golden values in request_bars test
The circus that is Alpaca messing with four year old historical data
continues. Similar to what we did to other tests in the not-so-distant
past, this change adjusts the data::v2::bars::tests::request_bars test
to use "new" golden values and to use integer conversions to make it
harder for certain clowns to make said test fail in the future.