Auto-derive Eq for various types implementing PartialEq
Clippy rightfully points out that various types for which we auto-derive
PartialEq could also implement Eq. The only concern we have doing so is
that it has the potential to negatively impact compile time of the
crate. However, given that we currently do not track that or have taken
any steps to optimize it, for now we just follow the suggestion and
implement Eq as well.