Allow clippy::result_large_err lint for api::v2::account_activities::Activity
The api::v2::account_activities::Activity is a large type, memory-wise.
The two convenience methods we have for conversion into a trade or
non-trade object have large Err variants because of this, which is
causing Clippy to complain. These are infrequently used methods and so
this is a trade-off we are willing to live with.
To that end, this change explicitly allows clippy::result_large_err lint
on both these conversion methods.