Adjust data::v2::last_quote to work for multiple symbols
In addition to the endpoint we use for querying the quote for a single
symbol, Alpaca has added another one that can be used to retrieve the
last quotes for multiple symbols. Given that doing so is basically
superior to the retrieval of a single quote, this change adjusts the
data::v2::last_quote module to work for multiple symbols.
The way we return quotes is in the form of <symbol, quote> tuples. An
alternative representation could embed the symbol into the Quote type
directly. There is no right answer as to which is better, but the
current thinking is that having the symbol stored outside of the type
makes it more universally usable and we are considering adjusting types
of other modules to follow suit (we already employ a mixture of the two
styles).