Deny warnings when generating documentation in CI
Our cargo doc invocation in CI is expected to fail when detecting
issues. However, given the way it is invoked currently, it is more
likely to emit warnings instead. Obviously, those may go undetected for
a long time.
To fix this issue, this change makes sure to set the RUSTDOCFLAGS in
such a way that cargo doc denies any warnings, converting them to errors
instead.