Add quantity_available field to api::v2::position::Position
A position object in Alpaca not only carries the total number of shares
as a property, but also the number of shares minus those not available
for trading at the moment because they are already covered by open
orders.
This change makes sure to expose this property via the newly added
quantity_available member to the api::v2::position::Position type.