Make more api::v2::order::ChangeReq members optional
Some of the members of the api::v2::order::ChangeReq type are
non-optional and at the same time have default representations. As such,
they do not technically have to be set, but at the same time the default
values may not actually be sensible. For example, the quantity is
default initialized to zero, which is not even an accepted input value
by the ALpaca API.
This change makes both the quanity and time_in_force members Option
objects. As such, they may now legitimately not be set, but the result
is conveyed correctly to the brokerage API.