Introduce vendored-openssl feature
The crate supports TLS secured connections via the hyper-tls crate.
That, in turn, uses the system's native opensl library (at least on
Linux). That can be a pain to work with in cross compilation contexts,
where additional system-dependent packages may now have to be installed.

The hyper-tls crate provides the vendored feature, which, a few layers
down, causes a vendored version of openssl to be used instead. That
provides more of a batteries-included experience.
With this change we introduce the vendored-openssl feature to the crate
which enables hyper-tls' vendored feature, which in turn causes the
library to use a vendored version of openssl.