Add proper CI job for building with minimum dependency versions
We have repeatedly been impacted by more recent versions of dependencies
(direct or transitive) requiring more recent versions of Rust in our
workflow that tests building with the minimum supported Rust version.
One reason for that is because we are a library crate and do not check
in our Cargo.lock file. That is unlikely to change. However, what we can
do is decouple ourselves from whatever the newest versions are and just
using the minimum versions available instead.
This change introduces a CI job that does precisely that, replacing the
existing one that builds with the minimum supported Rust version.