Report lack of account activities more prominently
Some of our account activity tests require two or more activities to be
present in order to work successfully. However, a lack of those is not
immediately obvious from the resulting test failures.
This change adjusts a test that retrieves account activity anyway and
checks for certain properties to also report if less than two activities
are currently present in a prominent and unambiguous way.