Switch to using new stream authentication request message format
When subscribing to order updates, for example, we were seeing a message
telling us that the format used for the authentication request was being
deprecated.
With this change we switch to adhering to the new format. To make that
possible, split the existing Request enumeration type in two, one using
a "data" container and the other not.
While at it, also make these two types private to the module they are
defined in, as they are not meant to be exposed in any way.