Use Num type for various data::v2::stream::{Bar,Quote,Trade} members
The Bar, Quote, and Trade types of the data::v2::stream module contain a
bunch of integer members pertaining trade sizes, volumes, and similar.
With the introduction of fractional trading there no longer is any
guarantee that reported values are indeed integers. That certainly is
not the case for Crypto streaming APIs.
With this change we adjust these members to be of the Num type instead,
which provides a greater deal of flexibility.