Enable support for custom v2 realtime data streaming URLs
Our v2 realtime data streaming URLs currently support two sources, which
indirectly influence the URL we connect to in the background: IEX and
SIP.
Having two hard coded values like this is great for usability (users
can't do much wrong), but it admittedly is not particularly open to
extension. Specifically, with crypto being all the rage these days and
it basically using nothing more than the v2 handshake and message format
(in version v1beta3), it would be nice for us to provide a way for users
to provide a URL to the crypto endpoint and, with that, provide users
with a way to stream crypto quotes, trades, and bars, without actually
hard coding details about the not-yet-stabilized Alpaca APIs into the
crate.
With this change we introduce a the new data::v2::stream::CustomUrl type
that enables such usage.

Closes: #56