Adjust publish workflow to create GitHub release
This change adjusts the publish workflow to create a proper GitHub
release as part of publishing the crate to crates.io. As part of cutting
such a release, a Git tag is created automatically as well. Having this
infrastructure in place reduces the number of manual steps that have to
be performed as part of the release process.