Adjust time stamp checks for trade retrieval tests
One may not have guessed, but the first trade on 2018-12-03T21:47:01
just started happening at a different millisecond! There must be
order. Better change that now in case someone relied on it, now that we
know when it really happened four years ago!!1one
Adjust our golden value check to ignore sub-second time stamp data
completely to not have to deal with this idiosyncrasy in the future.