Introduce NotPermitted variant to various data::v2 endpoint errors
It wouldn't be fun if the fucking documented API contract wouldn't be
broken every other month in some utterly braindead way. In this month's
version: HTTP status 403 is not used to convey the fact that data could
not be retrieved, possibly due to a missing data subscription.
With this change we introduce the new NotPermitted variant to the
various data::v2 endpoint error types to expose this condition.