Rename omnipresent AuthenticationFailed error to NotPermitted
The omnipresent AuthenticationFailed error variant, which used to map to
HTTP status 401 (Unauthorized) had not been hooked up for ages, after
Alpaca in their endless wisdom decided to move everything to using HTTP
status 403, causing ambiguity all over the place.
With this change we rename this variant to NotPermitted and map it to
HTTP status 403. As part of this work, we effectively replace/overlay
the recently introduced NotPermitted variants for the various data::v2
endpoint errors.