Rename api::v2::watchlist::Post type to Create
Our API request (and supporting) types are not exactly named
consistently. Some of them are named after the HTTP verb being used in
the underlying request, others are not. Some reflect the containing
module in their name, others don't.
This change is the first one in a series cleaning up this state of
affairs. Specifically, it renames the api::v2::watchlist::Post request
type to Create, which more accurately reflects its intended purpose.