Rename api::v2::account_config::Patch type to Change
In continuing our effort of improving the consistency of our type names,
this change renames the api::v2::account_config::Patch type to Change.
This verb more precisely reflects the action being performed and
decouples the request type naming from that of the underlying HTTP verb
(which users ultimately should not have to care about).