Remove api::v2::order::ChangeReqInit type
The api::v2::order::ChangeReq type is defacto default constructible,
because all attributes are optional. We introduced the ChangeReqInit
helper in anticipation of a potential addition of a
non-default-constructible member. However, such an addition would be
breaking the upstream API in a backwards compatible manner. While
certainly possible given Alpaca's track record of idiosyncrasies, we no
longer intend to cater to nonsense like that.
Remove the ChangeReqInit type and implement the Default trait for
ChangeReq directly to simplify the API surface.