Declare various types as non-exhaustive (again)
A while back we removed the non-exhaustive markers from a bunch of
types, to allow certain clients to construct objects of those types.
Having exhaustive struct types is a backwards compatibility hazard and
at least for some it is reasonable to anticipate future member
additions.
Since then we have found a way to straddle both worlds: that of allowing
certain users to create types while also making it clear that they are
doing so at their own risk and outside of SemVer guarantees. Basically,
we add a public but undocumented attribute that has to be explicitly
named for exhaustive construction. Anybody doing that is advised to pin
the version of the crate that they depend on.
This change adds such markers to and removes the #[non_exhaustive]
attribute from all struct types. For enums used in output positions, we
add #[non_exhaustive] where a future variant addition seems somewhat
likely. For those in input positions we add the #[non_exhaustive]
attribute pretty much unconditionally. Furthermore, a lot of enums used
in output positions already feature an Unknonwn variant. However, its
purpose is slightly different from the non-exhaustive attribute in that
it prevents parsing errors when an unknown variant is returned from
upstream. That is useful to not make the parsing of an otherwise valid
response fail prematurely and potentially cause mayhem, as we have seen
happen in the past.