Introduce api::v2::watchlist::CreateReqInit
The creation of a watch list always requires a name (so does the
updating, at least according to the documentation). Hence, it makes the
most sense to follow the "ReqInit" pattern that we employ elsewhere.
Doing so also allows us to make the CreateReq type non-exhaustive, which
will help with backwards compatibility.