Add api::v2::asset::Status::Unknown variant
With the api::v2::asset::Status being declared as non-exhaustive by now,
it seems logical to also support the deserialization of potentially
unknown variants, otherwise a deserialization failure could have
disastrous consequences when listing tradeable assets.
As such, this change adds an Unknown variant to the enum, similar to
what we have elsewhere.